Move the XBee_Custom_Packet class to its own file
